moves global components to service layer
