+++ frontend/src/Hooks/Store.tsx
@@ -1,6 +1,6 @@
 import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux';
 
-import type { AppDispatch, RootState } from '../store';
+import type { AppDispatch, RootState } from '../Services/ReduxStore/Store';
 
 // Use throughout your app instead of plain `useDispatch` and `useSelector`
 export const useAppDispatch = () => useDispatch<AppDispatch>();
